+++ implement/Pine.Core/LanguageServerProtocol/ServerCapabilities.cs
@@ -10,7 +10,10 @@ public record ServerCapabilities(
     TextDocumentSyncOptions? TextDocumentSync = null,
     ServerCapabilitiesWorkspace? Workspace = null,
     bool? HoverProvider = null,
-    CompletionOptions? CompletionProvider = null);
+    CompletionOptions? CompletionProvider = null,
+    bool? DeclarationProvider = null,
+    bool? DefinitionProvider = null,
+    bool? TypeDefinitionProvider = null);
 
 /// <summary>
 /// https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocumentSyncKind