+++ implement/pine/Program.cs
@@ -23,7 +23,7 @@ namespace ElmTime;
 
 public class Program
 {
-    public static string AppVersionId => "0.3.27";
+    public static string AppVersionId => "0.3.28";
 
     private static int AdminInterfaceDefaultPort => 4000;